+++ CHANGELOG.md
@@ -94,6 +94,7 @@ https://github.com/librespot-org/librespot
   Connect should use the 'filter-explicit-content' user attribute in the session.
 - [playback] Add metadata support via a `TrackChanged` event
 - [connect] Add `activate` and `load` functions to `Spirc`, allowing control over local connect sessions
+- [metadata] Add `Lyrics`
 
 ### Fixed